# Vectorize RetinaNet's postprocessing