--- a/torchvision/models/detection/retinanet.py
+++ b/torchvision/models/detection/retinanet.py
@@ -558,9 +558,8 @@ class RetinaNet(nn.Module):
             # split outputs per level
             split_head_outputs: Dict[str, List[Tensor]] = {}
             for k in head_outputs:
-                split_head_outputs[k] = [x.permute(1, 0, 2) for x in
-                                         head_outputs[k].permute(1, 0, 2).split_with_sizes(feature_sizes_per_level)]
-            split_anchors = [list(a.split_with_sizes(feature_sizes_per_level)) for a in anchors]
+                split_head_outputs[k] = list(head_outputs[k].split(feature_sizes_per_level, dim=1))
+            split_anchors = [list(a.split(feature_sizes_per_level)) for a in anchors]
 
             # compute the detections
             detections = self.postprocess_detections(split_head_outputs, split_anchors, images.image_sizes)