… test/expect/ModelTester.test_retinanet_resnet50_fpn_expect.pkl