@fmassa 2020-10-20
@hgaiser I think we could re-introduce it in a follow-up PR, but it would be good to have some more strict approaches for this.
For example, the `pop` from the dict is not ideal, as it changes the input dict in-place.